refactor: move labels/descriptions to bundle
