+++ src/main/resources/messages/codegpt.properties
@@ -3,6 +3,12 @@ notification.group.name=CodeGPT notification group
 action.generateCommitMessage.title=Generate Message
 action.generateCommitMessage.description=Generate commit message using OpenAI service
 action.generateCommitMessage.serviceWarning=Messages can only be generated with the OpenAI service.
+action.includeFilesInContext.title=Include In Context...
+action.includeFilesInContext.dialog.title=Include In Context
+action.includeFilesInContext.dialog.description=Choose the files that you wish to include in the final prompt
+action.includeFilesInContext.dialog.promptTemplate.label=Prompt template:
+action.includeFilesInContext.dialog.repeatableContext.label=Repeatable context:
+action.includeFilesInContext.dialog.restoreToDefaults.label=Restore to Defaults
 settings.displayName=CodeGPT: Settings
 settings.openaiQuotaExceeded=OpenAI quota exceeded.
 settingsConfigurable.displayName.label=Display name:
